Rename `rvc*` to `compressed*`
* Rename `rvc*` to `compressed*`, including:
    * `rvc.h` -> `compressed.h`
    * `rvc.c` -> `compressed.c`
    * `rvc_private.h` -> `compressed_private.h`
