+++ compressed_private.h
